+++ packages/netlify-cms-backend-github/src/__tests__/implementation.spec.js
@@ -185,7 +185,9 @@ describe('github backend implementation', () => {
         isModification: true,
         metaData: {
           branch: 'branch',
-          objects: { entry: { path: 'entry-path' }, files: [{ path: 'image.png', sha: 'sha' }] },
+          objects: {
+            entry: { path: 'entry-path', mediaFiles: [{ path: 'image.png', id: 'sha' }] },
+          },
         },
       };
       readUnpublishedBranchFile.mockResolvedValue(data);