+++ packages/netlify-cms-core/src/reducers/__tests__/collections.spec.js
@@ -4,8 +4,11 @@ import collections, {
   selectAllowDeletion,
   selectEntryPath,
   selectEntrySlug,
-  selectFieldsMediaFolders,
+  selectFieldsWithMediaFolders,
   selectMediaFolders,
+  getFieldsNames,
+  selectField,
+  updateFieldByKey,
 } from '../collections';
 import { FILES, FOLDER } from 'Constants/collectionTypes';
 
@@ -85,12 +88,12 @@ describe('collections', () => {
 
   describe('selectFieldsMediaFolders', () => {
     it('should return empty array for invalid collection', () => {
-      expect(selectFieldsMediaFolders(fromJS({}))).toEqual([]);
+      expect(selectFieldsWithMediaFolders(fromJS({}))).toEqual([]);
     });
 
     it('should return configs for folder collection', () => {
       expect(
-        selectFieldsMediaFolders(
+        selectFieldsWithMediaFolders(
           fromJS({
             folder: 'posts',
             fields: [
@@ -118,23 +121,43 @@ describe('collections', () => {
                   },
                 ],
               },
+              {
+                name: 'list_3',
+                types: [
+                  {
+                    name: 'list_3_type',
+                    media_folder: 'list_3_type_media_folder',
+                  },
+                ],
+              },
             ],
           }),
         ),
       ).toEqual([
-        'image_media_folder',
-        'body_media_folder',
-        'list_1_item_media_folder',
-        'list_2_item_media_folder',
+        fromJS({
+          name: 'image',
+          media_folder: 'image_media_folder',
+        }),
+        fromJS({ name: 'body', media_folder: 'body_media_folder' }),
+        fromJS({ name: 'list_1_item', media_folder: 'list_1_item_media_folder' }),
+        fromJS({
+          name: 'list_2_item',
+          media_folder: 'list_2_item_media_folder',
+        }),
+        fromJS({
+          name: 'list_3_type',
+          media_folder: 'list_3_type_media_folder',
+        }),
       ]);
     });
 
     it('should return configs for files collection', () => {
       expect(
-        selectFieldsMediaFolders(
+        selectFieldsWithMediaFolders(
           fromJS({
             files: [
               {
+                name: 'file1',
                 fields: [
                   {
                     name: 'image',
@@ -143,6 +166,7 @@ describe('collections', () => {
                 ],
               },
               {
+                name: 'file2',
                 fields: [
                   {
                     name: 'body',
@@ -151,6 +175,7 @@ describe('collections', () => {
                 ],
               },
               {
+                name: 'file3',
                 fields: [
                   {
                     name: 'list_1',
@@ -162,6 +187,7 @@ describe('collections', () => {
                 ],
               },
               {
+                name: 'file4',
                 fields: [
                   {
                     name: 'list_2',
@@ -170,18 +196,32 @@ describe('collections', () => {
                         name: 'list_2_item',
                         media_folder: 'list_2_item_media_folder',
                       },
+                      {
+                        name: 'list_3',
+                        types: [
+                          {
+                            name: 'list_3_type',
+                            media_folder: 'list_3_type_media_folder',
+                          },
+                        ],
+                      },
                     ],
                   },
                 ],
               },
             ],
           }),
+          'file4',
         ),
       ).toEqual([
-        'image_media_folder',
-        'body_media_folder',
-        'list_1_item_media_folder',
-        'list_2_item_media_folder',
+        fromJS({
+          name: 'list_2_item',
+          media_folder: 'list_2_item_media_folder',
+        }),
+        fromJS({
+          name: 'list_3_type',
+          media_folder: 'list_3_type_media_folder',
+        }),
       ]);
     });
   });
@@ -193,48 +233,258 @@ describe('collections', () => {
       sanitize_replacement: '-',
     };
 
-    const config = fromJS({ slug });
-    it('should return fields and collection folder', () => {
+    const config = fromJS({ slug, media_folder: '/static/img' });
+    it('should return fields and collection folders', () => {
       expect(
         selectMediaFolders(
           { config },
           fromJS({
             folder: 'posts',
-            media_folder: '/collection_media_folder',
+            media_folder: '{{media_folder}}/general/',
             fields: [
               {
                 name: 'image',
-                media_folder: '/image_media_folder',
+                media_folder: '{{media_folder}}/customers/',
+              },
+              {
+                name: 'list',
+                types: [{ name: 'widget', media_folder: '{{media_folder}}/widgets' }],
               },
             ],
           }),
-          fromJS({ slug: 'name', path: 'src/post/post1.md' }),
+          fromJS({ slug: 'name', path: 'src/post/post1.md', data: {} }),
         ),
-      ).toEqual(['collection_media_folder', 'image_media_folder']);
+      ).toEqual([
+        'static/img/general',
+        'static/img/general/customers',
+        'static/img/general/widgets',
+      ]);
     });
 
-    it('should return fields and collection folder', () => {
+    it('should return fields, file and collection folders', () => {
       expect(
         selectMediaFolders(
           { config },
           fromJS({
+            media_folder: '{{media_folder}}/general/',
             files: [
               {
                 name: 'name',
                 file: 'src/post/post1.md',
-                media_folder: '/file_media_folder',
+                media_folder: '{{media_folder}}/customers/',
                 fields: [
                   {
                     name: 'image',
-                    media_folder: '/image_media_folder',
+                    media_folder: '{{media_folder}}/logos/',
+                  },
+                  {
+                    name: 'list',
+                    types: [{ name: 'widget', media_folder: '{{media_folder}}/widgets' }],
                   },
                 ],
               },
             ],
           }),
-          fromJS({ slug: 'name', path: 'src/post/post1.md' }),
+          fromJS({ slug: 'name', path: 'src/post/post1.md', data: {} }),
         ),
-      ).toEqual(['file_media_folder', 'image_media_folder']);
+      ).toEqual([
+        'static/img/general',
+        'static/img/general/customers',
+        'static/img/general/customers/logos',
+        'static/img/general/customers/widgets',
+      ]);
+    });
+  });
+
+  describe('getFieldsNames', () => {
+    it('should get flat fields names', () => {
+      const collection = fromJS({
+        fields: [{ name: 'en' }, { name: 'es' }],
+      });
+      expect(getFieldsNames(collection.get('fields').toArray())).toEqual(['en', 'es']);
+    });
+
+    it('should get nested fields names', () => {
+      const collection = fromJS({
+        fields: [
+          { name: 'en', fields: [{ name: 'title' }, { name: 'body' }] },
+          { name: 'es', fields: [{ name: 'title' }, { name: 'body' }] },
+          { name: 'it', field: { name: 'title', fields: [{ name: 'subTitle' }] } },
+          {
+            name: 'fr',
+            fields: [{ name: 'title', widget: 'list', types: [{ name: 'variableType' }] }],
+          },
+        ],
+      });
+      expect(getFieldsNames(collection.get('fields').toArray())).toEqual([
+        'en',
+        'es',
+        'it',
+        'fr',
+        'en.title',
+        'en.body',
+        'es.title',
+        'es.body',
+        'it.title',
+        'it.title.subTitle',
+        'fr.title',
+        'fr.title.variableType',
+      ]);
+    });
+  });
+
+  describe('selectField', () => {
+    it('should return top field by key', () => {
+      const collection = fromJS({
+        fields: [{ name: 'en' }, { name: 'es' }],
+      });
+      expect(selectField(collection, 'en')).toBe(collection.get('fields').get(0));
+    });
+
+    it('should return nested field by key', () => {
+      const collection = fromJS({
+        fields: [
+          { name: 'en', fields: [{ name: 'title' }, { name: 'body' }] },
+          { name: 'es', fields: [{ name: 'title' }, { name: 'body' }] },
+          { name: 'it', field: { name: 'title', fields: [{ name: 'subTitle' }] } },
+          {
+            name: 'fr',
+            fields: [{ name: 'title', widget: 'list', types: [{ name: 'variableType' }] }],
+          },
+        ],
+      });
+
+      expect(selectField(collection, 'en.title')).toBe(
+        collection
+          .get('fields')
+          .get(0)
+          .get('fields')
+          .get(0),
+      );
+
+      expect(selectField(collection, 'it.title.subTitle')).toBe(
+        collection
+          .get('fields')
+          .get(2)
+          .get('field')
+          .get('fields')
+          .get(0),
+      );
+
+      expect(selectField(collection, 'fr.title.variableType')).toBe(
+        collection
+          .get('fields')
+          .get(3)
+          .get('fields')
+          .get(0)
+          .get('types')
+          .get(0),
+      );
+    });
+  });
+
+  describe('updateFieldByKey', () => {
+    it('should update field by key', () => {
+      const collection = fromJS({
+        fields: [
+          { name: 'title' },
+          { name: 'image' },
+          {
+            name: 'object',
+            fields: [{ name: 'title' }, { name: 'gallery', fields: [{ name: 'image' }] }],
+          },
+          { name: 'list', field: { name: 'image' } },
+          { name: 'body' },
+          { name: 'widgetList', types: [{ name: 'widget' }] },
+        ],
+      });
+
+      const updater = field => field.set('default', 'default');
+
+      expect(updateFieldByKey(collection, 'non-existent', updater)).toBe(collection);
+      expect(updateFieldByKey(collection, 'title', updater)).toEqual(
+        fromJS({
+          fields: [
+            { name: 'title', default: 'default' },
+            { name: 'image' },
+            {
+              name: 'object',
+              fields: [{ name: 'title' }, { name: 'gallery', fields: [{ name: 'image' }] }],
+            },
+            { name: 'list', field: { name: 'image' } },
+            { name: 'body' },
+            { name: 'widgetList', types: [{ name: 'widget' }] },
+          ],
+        }),
+      );
+      expect(updateFieldByKey(collection, 'object.title', updater)).toEqual(
+        fromJS({
+          fields: [
+            { name: 'title' },
+            { name: 'image' },
+            {
+              name: 'object',
+              fields: [
+                { name: 'title', default: 'default' },
+                { name: 'gallery', fields: [{ name: 'image' }] },
+              ],
+            },
+            { name: 'list', field: { name: 'image' } },
+            { name: 'body' },
+            { name: 'widgetList', types: [{ name: 'widget' }] },
+          ],
+        }),
+      );
+
+      expect(updateFieldByKey(collection, 'object.gallery.image', updater)).toEqual(
+        fromJS({
+          fields: [
+            { name: 'title' },
+            { name: 'image' },
+            {
+              name: 'object',
+              fields: [
+                { name: 'title' },
+                { name: 'gallery', fields: [{ name: 'image', default: 'default' }] },
+              ],
+            },
+            { name: 'list', field: { name: 'image' } },
+            { name: 'body' },
+            { name: 'widgetList', types: [{ name: 'widget' }] },
+          ],
+        }),
+      );
+      expect(updateFieldByKey(collection, 'list.image', updater)).toEqual(
+        fromJS({
+          fields: [
+            { name: 'title' },
+            { name: 'image' },
+            {
+              name: 'object',
+              fields: [{ name: 'title' }, { name: 'gallery', fields: [{ name: 'image' }] }],
+            },
+            { name: 'list', field: { name: 'image', default: 'default' } },
+            { name: 'body' },
+            { name: 'widgetList', types: [{ name: 'widget' }] },
+          ],
+        }),
+      );
+
+      expect(updateFieldByKey(collection, 'widgetList.widget', updater)).toEqual(
+        fromJS({
+          fields: [
+            { name: 'title' },
+            { name: 'image' },
+            {
+              name: 'object',
+              fields: [{ name: 'title' }, { name: 'gallery', fields: [{ name: 'image' }] }],
+            },
+            { name: 'list', field: { name: 'image' } },
+            { name: 'body' },
+            { name: 'widgetList', types: [{ name: 'widget', default: 'default' }] },
+          ],
+        }),
+      );
     });
   });
 });