+++ packages/netlify-cms-locales/src/cs/index.js
@@ -3,6 +3,9 @@ const cs = {
     login: 'Přihlásit',
     loggingIn: 'Přihlašování...',
     loginWithNetlifyIdentity: 'Přihlásit pomocí Netlify Identity',
+    loginWithBitbucket: 'Přihlásit pomocí Bitbucket',
+    loginWithGitHub: 'Přihlásit pomocí GitHub',
+    loginWithGitLab: 'Přihlásit pomocí GitLab',
     errors: {
       email: 'Vyplňte emailovou adresu.',
       password: 'Vyplňte heslo.',
@@ -45,6 +48,11 @@ const cs = {
     },
   },
   editor: {
+    editorControl: {
+      field: {
+        optional: 'volitelný',
+      },
+    },
     editorControlPane: {
       widget: {
         required: '%{fieldLabel} je povinný.',
@@ -101,6 +109,16 @@ const cs = {
       deployButtonLabel: 'Zobrazit na webu',
     },
     editorWidgets: {
+      image: {
+        choose: 'vyberte obrázek',
+        chooseDifferent: 'vyberte jiný obrázek',
+        remove: 'odstranit obrázek',
+      },
+      file: {
+        choose: 'vyberte soubor',
+        chooseDifferent: 'vyberte jiný soubor',
+        remove: 'odebrat soubor',
+      },
       unknownControl: {
         noControl: "Žádné ovládání pro widget '%{widget}'.",
       },