+++ packages/netlify-cms-locales/src/zh_Hant/index.js
@@ -3,6 +3,9 @@ const zh_Hant = {
     login: '登入',
     loggingIn: '正在登入...',
     loginWithNetlifyIdentity: '使用你的 Netlify 帳號來進行登入',
+    loginWithBitbucket: '使用你的 Bitbucket 帳號來進行登入',
+    loginWithGitHub: '使用你的 GitHub 帳號來進行登入',
+    loginWithGitLab: '使用你的 GitLab 帳號來進行登入',
     errors: {
       email: '請確認你已經輸入你的電子郵件。',
       password: '請輸入你的密碼。',
@@ -44,6 +47,11 @@ const zh_Hant = {
     },
   },
   editor: {
+    editorControl: {
+      field: {
+        optional: '可选的',
+      },
+    },
     editorControlPane: {
       widget: {
         required: '%{fieldLabel} 是必須的。',
@@ -101,6 +109,16 @@ const zh_Hant = {
       deployButtonLabel: '觀看已發布的內容',
     },
     editorWidgets: {
+      image: {
+        choose: '选择一张图片',
+        chooseDifferent: '选择其他图片',
+        remove: '移除图片',
+      },
+      file: {
+        choose: '选择一个文件',
+        chooseDifferent: '选择其他文件',
+        remove: '删除文件',
+      },
       unknownControl: {
         noControl: "無法控制元件： '%{widget}'.",
       },