+++ packages/netlify-cms-proxy-server/src/index.ts
@@ -11,7 +11,7 @@ const port = process.env.PORT || 8081;
 (async () => {
   app.use(morgan('combined'));
   app.use(cors());
-  app.use(express.json());
+  app.use(express.json({ limit: '50mb' }));
 
   try {
     const mode = process.env.MODE || 'fs';